Refactored 2-step code to work with Dev Portal too
Moved all 2 step code into `two_step_client.rb`
