+++ spaceship/lib/spaceship/portal/portal_client.rb
@@ -42,7 +42,7 @@ module Spaceship
         # User Credentials are wrong
         raise InvalidUserCredentialsError.new, "Invalid username and password combination. Used '#{user}' as the username."
       elsif (response.body || "").include?("Verify your identity")
-        raise "spaceship / fastlane doesn't support 2 step enabled accounts yet. Please temporary disable 2 step verification until spaceship was updated."
+        return send_itc_login_request(user, password) # 2 step JSON API is only offered on iTC
       end
 
       case response.status
