+++ spaceship/lib/spaceship/tunes/tunes_client.rb
@@ -116,167 +116,9 @@ module Spaceship
       end
     end
 
-    def service_key
-      return @service_key if @service_key
-      # We need a service key from a JS file to properly auth
-      js = request(:get, "https://itunesconnect.apple.com/itc/static-resources/controllers/login_cntrl.js")
-      @service_key ||= js.body.match(/itcServiceKey = '(.*)'/)[1]
-    end
-
     def send_login_request(user, password)
       clear_user_cached_data
-      data = {
-        accountName: user,
-        password: password,
-        rememberMe: true
-      }
-
-      begin
-        response = request(:post) do |req|
-          req.url "https://idmsa.apple.com/appleauth/auth/signin?widgetKey=#{service_key}"
-          req.body = data.to_json
-          req.headers['Content-Type'] = 'application/json'
-          req.headers['X-Requested-With'] = 'XMLHttpRequest'
-          req.headers['Accept'] = 'application/json, text/javascript'
-        end
-      rescue UnauthorizedAccessError
-        raise InvalidUserCredentialsError.new, "Invalid username and password combination. Used '#{user}' as the username."
-      end
-
-      # get woinst, wois, and itctx cookie values
-      request(:get, "https://itunesconnect.apple.com/WebObjects/iTunesConnect.woa/wa/route?noext")
-      request(:get, "https://itunesconnect.apple.com/WebObjects/iTunesConnect.woa")
-
-      case response.status
-      when 403
-        raise InvalidUserCredentialsError.new, "Invalid username and password combination. Used '#{user}' as the username."
-      when 200
-        return response
-      else
-        if response["Location"] == "/auth" # redirect to 2 step auth page
-          if load_session_from_file
-            return true
-          end
-
-          handle_two_step(response)
-          return true
-        elsif (response.body || "").include?('invalid="true"')
-          # User Credentials are wrong
-          raise InvalidUserCredentialsError.new, "Invalid username and password combination. Used '#{user}' as the username."
-        elsif (response['Set-Cookie'] || "").include?("itctx")
-          raise "Looks like your Apple ID is not enabled for iTunes Connect, make sure to be able to login online"
-        else
-          info = [response.body, response['Set-Cookie']]
-          raise ITunesConnectError.new, info.join("\n")
-        end
-      end
-    end
-
-    # Only needed for 2 step
-    def load_session_from_file
-      if File.exist?(persistent_cookie_path)
-        @cookie.load(persistent_cookie_path)
-        if self.team_id.to_s.length > 0 # this will send a request that will fail if the cookie is invalid
-          return true
-        else
-          File.delete(persistent_cookie_path) # session is already invalid
-        end
-      end
-      return false
-    end
-
-    def handle_two_step(response)
-      @x_apple_web_session_token = response["x-apple-web-session-token"]
-      @scnt = response["scnt"]
-
-      r = request(:get) do |req|
-        req.url "https://idmsa.apple.com/appleauth/auth"
-        req.headers["scnt"] = @scnt
-        req.headers["X-Apple-Web-Session-Token"] = @x_apple_web_session_token
-        req.headers["Accept"] = "application/json"
-      end
-
-      if r.body.kind_of?(Hash) && r.body["trustedDevices"].kind_of?(Array)
-        if r.body.fetch("securityCode", {})["tooManyCodesLock"].to_s.length > 0
-          raise ITunesConnectError.new, "Too many verification codes have been sent. Enter the last code you received, use one of your devices, or try again later."
-        end
-
-        old_client = (begin
-                        Tunes::RecoveryDevice.client
-                      rescue
-                        nil # since client might be nil, which raises an exception
-                      end)
-        Tunes::RecoveryDevice.client = self # temporary set it as it's required by the factory method
-        devices = r.body["trustedDevices"].collect do |current|
-          Tunes::RecoveryDevice.factory(current)
-        end
-        Tunes::RecoveryDevice.client = old_client
-
-        puts "Two Step Verification for account '#{self.user}' is enabled"
-        puts "Please select a device to verify your identity"
-        available = devices.collect do |c|
-          "#{c.name}\t#{c.model_name || 'SMS'}\t(#{c.device_id})"
-        end
-        result = choose(*available)
-        device_id = result.match(/.*\t.*\t\((.*)\)/)[1]
-        select_device(r, device_id)
-      else
-        raise "Invalid 2 step response #{r.body}"
-      end
-    end
-
-    def select_device(r, device_id)
-      # Request Token
-      r = request(:put) do |req|
-        req.url "https://idmsa.apple.com/appleauth/auth/verify/device/#{device_id}/securitycode"
-        req.headers["Accept"] = "application/json"
-        req.headers["scnt"] = @scnt
-        req.headers["X-Apple-Web-Session-Token"] = @x_apple_web_session_token
-      end
-      handle_itc_response(r.body)
-
-      puts "Successfully requested notification"
-      code = ask("Please enter the 4 digit code: ")
-      puts "Requesting session..."
-
-      # Send token back to server to get a valid session
-      r = request(:post) do |req|
-        req.url "https://idmsa.apple.com/appleauth/auth/verify/device/#{device_id}/securitycode"
-        req.headers["Accept"] = "application/json"
-        req.headers["scnt"] = @scnt
-        req.headers["X-Apple-Web-Session-Token"] = @x_apple_web_session_token
-        req.body = { "code" => code.to_s }.to_json
-        req.headers['Content-Type'] = 'application/json'
-      end
-      handle_itc_response(r.body) # this will fail if the code is invalid
-      # If it works, r.body is actually nil
-      raise "Could not login" unless r.headers["set-cookie"].include?("myacinfo")
-
-      # Tell iTC that we are trustworthy (obviously)
-      # This will update our local cookies to something new
-      # They probably have a longer time to live than the other poor cookies
-      # Changed Keys
-      # - myacinfo
-      # - DES5c148586dfd451e55afb0175f62418f91
-
-      request(:get) do |req|
-        req.url "https://idmsa.apple.com/appleauth/auth/2sv/trust"
-        req.headers["scnt"] = @scnt
-        req.headers["X-Apple-Web-Session-Token"] = @x_apple_web_session_token
-      end
-
-      # We want to store the myacinfo cookie
-      @cookie.save(persistent_cookie_path, :yaml, session: true)
-      # really important to specify the session
-      # otherwise myacinfo and more won't be stored
-
-      return true
-    end
-
-    def persistent_cookie_path
-      path = File.expand_path(File.join("~", ".spaceship", self.user, "cookie"))
-      FileUtils.mkdir_p(File.expand_path("..", path))
-      return path
+      send_itc_login_request(user, password)
     end
 
     # rubocop:disable Metrics/CyclomaticComplexity
